Handle batch_slice serialization in Checkpoint
